Added Fragments to the project
Fixed comments

Comments fixed

Fixed mauri's comments

+++ app/src/main/java/com/mobile/tandil/javabaseproject/mvp/model/ParkingModel.java
@@ -4,16 +4,16 @@ import com.mobile.tandil.javabaseproject.mvp.contract.ParkingContract;
 
 public class ParkingModel implements ParkingContract.Model {
 
-    private static final int DEFAULT_VALUE = -1;
-    private int availableParkingSpots = DEFAULT_VALUE;
+    private static final long DEFAULT_VALUE = -1;
+    private long availableParkingSpots = DEFAULT_VALUE;
 
     @Override
-    public void setAvailableParkingSpots(int availableParkingSpots){
+    public void setAvailableParkingSpots(Long availableParkingSpots) {
         this.availableParkingSpots = availableParkingSpots;
     }
 
     @Override
-    public int getAvailableParkingSpots(){
+    public long getAvailableParkingSpots() {
         return availableParkingSpots;
     }
 }
